* Update cart/order functions: - Old: +Only can add 1 product at time.-> must add product to cart multiple time if customer want to buy more than 1 unit/product. +When place order, the product stock of product only decreased by 1. +Did not record order date. -New: +Customer can choose how many unit they want for each product they add to cart. +The product stock is decreased correctly (new_stock = old_stock - number_of_unit_in_cart) +Show order date.

+++ NMarket/src/main/java/com/Nkosopa/NMarket/DTO/Product/ProductDTO.java
@@ -17,7 +17,7 @@ public class ProductDTO extends BaseDTO<ProductDTO> {
 
     private String sku;
 
-    private int stock;
+    private Long stock;
 
     private Long price;
